+++ tools/release.sh
@@ -223,6 +223,7 @@ PAYLOAD_FILES=(
     cinder-home/dist/stable/cinder-voltable
     cinder-home/dist/stable/cinder-battery
     cinder-home/dist/stable/cinder-signature.sh
+    cinder-home/dist/stable/libcinder_mono.so
     cinder-home/dist/stable/cinder_components.conf
     cinder-home/dist/stable/cinder_home_install.upg
     cinder-home/dist/stable/cinder_home_uninstall.upg